# komga-rust/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Komga Rust API",
    "version": "0.1.0",
    "description": "Rust implementation of Komga media server for comics/manga. REST API for managing libraries, series, books, read progress, readlists, collections, and more."
  },
  "servers": [
    {
      "url": "/",
      "description": "Default server"
    }
  ],
  "paths": {
...
    "/api/v1/readlists": {
      "get": {
        "tags": ["ReadLists"],
        "summary": "List all readlists",
        "operationId": "listReadlists",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of readlists",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReadListPageDto" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["ReadLists"],
        "summary": "Create a readlist",
        "operationId": "createReadlist",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateReadListRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Readlist created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReadListDto" }
              }
            }
          }
        }
      }
    },
    "/api/v1/readlists/{id}": {
      "get": {
        "tags": ["ReadLists"],
        "summary": "Get readlist by ID",
        "operationId": "getReadlist",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Readlist details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReadListDto" }
              }
            }
          },
          "404": { "description": "Readlist not found" }
        }
      },
      "patch": {
        "tags": ["ReadLists"],
        "summary": "Update readlist",
        "operationId": "updateReadlist",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateReadListRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Readlist updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReadListDto" }
              }
            }
          },
          "404": { "description": "Readlist not found" }
        }
      },
      "delete": {
        "tags": ["ReadLists"],
        "summary": "Delete readlist",
        "operationId": "deleteReadlist",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Readlist deleted" },
          "404": { "description": "Readlist not found" }
        }
      }
    },
    "/api/v1/readlists/{id}/books": {
      "get": {
        "tags": ["ReadLists"],
        "summary": "List books in a readlist",
        "operationId": "getReadlistBooks",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of books",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BookPageDto" }
              }
            }
          },
          "404": { "description": "Readlist not found" }
        }
      }
    },
    "/api/v1/readlists/{id}/books/{bookId}/next": {
      "get": {
        "tags": ["ReadLists"],
        "summary": "Get next book in readlist",
        "operationId": "getReadlistNextBook",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "bookId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Next book (or null)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BookDto" }
              }
            }
          }
        }
      }
    },
    "/api/v1/readlists/{id}/books/{bookId}/previous": {
      "get": {
        "tags": ["ReadLists"],
        "summary": "Get previous book in readlist",
        "operationId": "getReadlistPreviousBook",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "bookId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Previous book (or null)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BookDto" }
              }
            }
          }
        }
      }
    },
    "/api/v1/readlists/{id}/file": {
      "get": {
        "tags": ["ReadLists"],
        "summary": "Download readlist as ZIP archive",
        "operationId": "getReadlistFile",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "ZIP file containing all books in readlist",
            "content": {
              "application/zip": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          }
        }
      }
    },
    "/api/v1/readlists/{id}/read-progress/tachiyomi": {
      "get": {
        "tags": ["ReadLists"],
        "summary": "Get Tachiyomi sync progress for readlist",
        "operationId": "getReadlistTachiyomiProgress",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Tachiyomi progress data",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["ReadLists"],
        "summary": "Update Tachiyomi sync progress for readlist",
        "operationId": "updateReadlistTachiyomiProgress",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "readChapters": {
                    "type": "array",
                    "items": { "type": "string" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Progress updated" }
        }
      }
    },
    "/api/v1/readlists/{id}/thumbnails": {
      "get": {
        "tags": ["ReadLists"],
        "summary": "List readlist thumbnails",
        "operationId": "listReadlistThumbnails",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "List of thumbnails",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "object" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["ReadLists"],
        "summary": "Create readlist thumbnail",
        "operationId": "createReadlistThumbnail",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Thumbnail created" }
        }
      }
    },
    "/api/v1/readlists/{id}/thumbnails/{thumbnailId}": {
      "get": {
        "tags": ["ReadLists"],
        "summary": "Get readlist thumbnail by ID",
        "operationId": "getReadlistThumbnailById",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "thumbnailId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Thumbnail image",
            "content": {
              "image/*": { "schema": { "type": "string", "format": "binary" } }
            }
          },
          "404": { "description": "Thumbnail not found" }
        }
      },
      "delete": {
        "tags": ["ReadLists"],
        "summary": "Delete readlist thumbnail",
        "operationId": "deleteReadlistThumbnail",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "thumbnailId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Thumbnail deleted" }
        }
      }
    },
    "/api/v1/readlists/{id}/thumbnails/{thumbnailId}/selected": {
      "put": {
        "tags": ["ReadLists"],
        "summary": "Mark readlist thumbnail as selected",
        "operationId": "markReadlistThumbnailSelected",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "thumbnailId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "202": { "description": "Thumbnail marked as selected" }
        }
      }
    },
    "/api/v1/readlists/{id}/thumbnail": {
      "get": {
        "tags": ["ReadLists"],
        "summary": "Get selected readlist thumbnail",
        "operationId": "getReadlistThumbnail",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Thumbnail image",
            "content": {
              "image/*": { "schema": { "type": "string", "format": "binary" } }
            }
          },
          "404": { "description": "Thumbnail not found" }
        }
      }
    },
    "/api/v1/readlists/match/comicrack": {
      "post": {
        "tags": ["ReadLists"],
        "summary": "Match ComicRack readlist",
        "operationId": "matchComicrackReadlist",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created readlist",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReadListDto" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "url": { "type": "string" },
          "seriesId": { "type": "string", "format": "uuid" },
          "seriesTitle": { "type": "string" },
          "number": { "type": "integer" },
          "oneshot": { "type": "boolean" },
          "fileHash": { "type": "string" },
          "sizeBytes": { "type": "integer", "format": "int64" },
          "size": { "type": "string" },
          "libraryId": { "type": "string", "format": "uuid" },
          "created": { "type": "string", "format": "date-time" },
          "lastModified": { "type": "string", "format": "date-time" },
          "fileLastModified": { "type": "string", "format": "date-time", "nullable": true },
          "deleted": { "type": "boolean" },
          "media": { "$ref": "#/components/schemas/MediaDto" },
          "metadata": { "$ref": "#/components/schemas/BookMetadataDto" },
          "readProgress": { "$ref": "#/components/schemas/ReadProgressDto" }
        }
      },
      "BookPageDto": {
        "type": "object",
        "properties": {
          "content": { "type": "array", "items": { "$ref": "#/components/schemas/BookDto" } },
          "totalElements": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "number": { "type": "integer" },
          "size": { "type": "integer" },
          "empty": { "type": "boolean" },
          "first": { "type": "boolean" },
          "last": { "type": "boolean" },
          "numberOfElements": { "type": "integer" }
        }
      },
...
      "MediaDto": {
        "type": "object",
        "properties": {
          "fileSize": { "type": "integer", "format": "int64" },
          "mediaType": { "type": "string" },
          "width": { "type": "integer", "nullable": true },
          "height": { "type": "integer", "nullable": true },
          "pagesCount": { "type": "integer" },
          "status": { "type": "string" }
        }
      },
...
      "ReadProgressDto": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "completed": { "type": "boolean" },
          "readDate": { "type": "string", "format": "date-time", "nullable": true },
          "deviceId": { "type": "string", "nullable": true },
          "deviceName": { "type": "string", "nullable": true },
          "created": { "type": "string", "format": "date-time", "nullable": true },
          "lastModified": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
      "ReadListDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "summary": { "type": "string" },
          "ordered": { "type": "boolean" },
          "filtered": { "type": "boolean" },
          "bookIds": { "type": "array", "items": { "type": "string", "format": "uuid" } },
          "createdDate": { "type": "string", "format": "date-time" },
          "lastModifiedDate": { "type": "string", "format": "date-time" }
        }
      },
      "ReadListPageDto": {
        "type": "object",
        "properties": {
          "content": { "type": "array", "items": { "$ref": "#/components/schemas/ReadListDto" } },
          "totalElements": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "number": { "type": "integer" },
          "size": { "type": "integer" },
          "empty": { "type": "boolean" },
          "first": { "type": "boolean" },
          "last": { "type": "boolean" },
          "numberOfElements": { "type": "integer" }
        }
      },
      "CreateReadListRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string" },
          "summary": { "type": "string" },
          "ordered": { "type": "boolean" },
          "bookIds": { "type": "array", "items": { "type": "string", "format": "uuid" } }
        }
      },
      "UpdateReadListRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "summary": { "type": "string" },
          "ordered": { "type": "boolean" },
          "bookIds": { "type": "array", "items": { "type": "string", "format": "uuid" } }
        }
      },
...
    }
  },
...
}